requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1;python_version<"3.9"
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-cors-headers==3.13.0
django-crum==0.7.9
django-filter==22.1
django-import-export==3.0.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
fonttools==4.34.4
future==0.18.2
html5lib==1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
# mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
Pillow==9.1.1
protobuf==4.21.1
psycopg2-binary
pycparser==2.21
pydyf==0.2.0
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pyparsing==3.0.9
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.10
weasyprint==56.1
webencodings==0.5.1
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.1
gunicorn
